fix src dir
