+++ backend/src/main/java/com/example/intervialion/controller/CategoryController.java
@@ -1,0 +1,31 @@
+package com.example.intervialion.controller;
+
+import com.example.intervialion.model.Category;
+import com.example.intervialion.service.CategoryRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
+
+@RestController
+@RequestMapping("api/category")
+@CrossOrigin(origins = "*")
+public class CategoryController {
+    @Autowired
+    private CategoryRepository categoryRepository;
+
+    @GetMapping("/GetCategories")
+    public ResponseEntity<List<Category>> getCategories() {
+        List<Category> categories = categoryRepository.findAll();
+        if (categories.isEmpty()) {
+            return ResponseEntity.noContent().build();
+        }
+        return new ResponseEntity<>(categories, HttpStatus.OK);
+    }
+
+
+}
+
