+++ backend/src/main/java/com/example/intervialion/model/SubCategory.java
@@ -1,0 +1,57 @@
+package com.example.intervialion.model;
+import com.fasterxml.jackson.annotation.JsonBackReference;
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.fasterxml.jackson.annotation.JsonManagedReference;
+import jakarta.persistence.*;
+import java.util.List;
+
+@Entity
+public class SubCategory {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    private String name;
+
+    @ManyToOne
+    @JoinColumn(name = "category_id")
+    @JsonBackReference // מונע חזרה אינסופית מהקטגוריה
+    private Category category;
+
+    @OneToMany(mappedBy = "subCategory", cascade = CascadeType.ALL, fetch = FetchType.LAZY)
+    @JsonManagedReference // טוען את המוצרים תחת תת הקטגוריה
+    private List<Product> products;
+    public SubCategory() {
+    }
+    public SubCategory(Long id, String name, Category category, List<Product> products) {
+        this.id = id;
+        this.name = name;
+        this.category = category;
+        this.products = products;
+    }
+    public Long getId() {
+        return id;
+    }
+    public void setId(Long id) {
+        this.id = id;
+    }
+    public String getName() {
+        return name;
+    }
+    public void setName(String name) {
+        this.name = name;
+    }
+    public Category getCategory() {
+        return category;
+    }
+    public void setCategory(Category category) {
+        this.category = category;
+    }
+    public List<Product> getProducts() {
+        return products;
+    }
+    public void setProducts(List<Product> products) {
+        this.products = products;
+    }
+}
+
