+++ backend/src/main/java/com/example/intervialion/service/MangerRepository.java
@@ -1,0 +1,6 @@
+package com.example.intervialion.service;
+import com.example.intervialion.model.Manger;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MangerRepository extends JpaRepository<Manger, Long> {
+}
