+++ backend/src/main/java/com/example/intervialion/service/ProductRepository.java
@@ -1,0 +1,9 @@
+package com.example.intervialion.service;
+import com.example.intervialion.model.Product;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+
+public interface ProductRepository extends JpaRepository<Product, Long> {
+    List<Product> findBySubCategoryId(Long subCategoryId);
+}
