--- backend/src/src/main/java/com/example/intervialion/controller/ProductController.java
@@ -1,39 +1,0 @@
-package com.example.intervialion.controller;
-
-import com.example.intervialion.model.Product;
-import com.example.intervialion.service.ProductRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
-@Configuration
-
-@RestController
-@RequestMapping("api/product")
-@CrossOrigin(origins = "*")
-public class ProductController {
-    @Autowired
-    private ProductRepository productRepository;
-
-    @GetMapping("/GetProducts")
-    public ResponseEntity<List<Product>> getProducts() {
-        List<Product> products = productRepository.findAll();
-        if (products.isEmpty()) {
-            return ResponseEntity.noContent().build();
-        }
-        return new ResponseEntity<>(products, HttpStatus.OK);
-    }
-
-    @GetMapping("/GetProductsBySubCategory/{subCategoryId}")
-    public ResponseEntity<List<Product>> getProductsBySubCategory(@PathVariable Long subCategoryId) {
-        List<Product> products = productRepository.findBySubCategoryId(subCategoryId);
-        if (products.isEmpty()) {
-            return ResponseEntity.noContent().build();
-        }
-        return new ResponseEntity<>(products, HttpStatus.OK);
-    }
-
-}
