--- backend/src/src/main/java/com/example/intervialion/model/Category.java
@@ -1,64 +1,0 @@
-package com.example.intervialion.model;
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import com.fasterxml.jackson.annotation.JsonManagedReference;
-import jakarta.persistence.*;
-
-import java.util.List;
-@Entity
-public class Category {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
-    private String name;
-    private String description;
-    @OneToMany(mappedBy = "category", cascade = CascadeType.ALL, fetch = FetchType.LAZY)
-    @JsonManagedReference // טוען את תתי הקטגוריות מבלי להיכנס לסיבוב אינסופי
-    private List<SubCategory> subCategories;
-
-
-
-    public Category() {
-    }
-
-    public Category(long id, String name, String description, List<SubCategory> subCategories) {
-        this.id = id;
-        this.name = name;
-        this.description = description;
-        this.subCategories = subCategories;
-
-    }
-
-    public long getId() {
-        return id;
-    }
-
-    public void setId(long id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-    public List<SubCategory> getSubCategories() {
-        return subCategories;
-    }
-
-    public void setSubCategories(List<SubCategory> subCategories) {
-        this.subCategories = subCategories;
-    }
-
-
-}
