--- backend/src/src/main/java/com/example/intervialion/service/CategoryRepository.java
@@ -1,7 +1,0 @@
-package com.example.intervialion.service;
-import com.example.intervialion.model.Category;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CategoryRepository extends JpaRepository <Category, Long>{
-
-}
