--- backend/src/src/main/java/com/example/intervialion/service/UserRepository.java
@@ -1,10 +1,0 @@
-package com.example.intervialion.service;
-import com.example.intervialion.model.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-
-}
-
-
-
